backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
openai
chromadb
tiktoken
pypdf
python-docx
python-multipart
requests